move auth to impl/auth and formatting
