+++ src/classes/Authenticator.js
@@ -1,10 +1,11 @@
-import decodeKey from '../authenticator/decodeKey';
-import encodeKey from '../authenticator/encodeKey';
-import keyuri from '../authenticator/keyuri';
-import qrcode from '../authenticator/qrcode';
-import secretKey from '../authenticator/secretKey';
-import token from '../authenticator/token';
+import decodeKey from '../impl/authenticator/decodeKey';
+import encodeKey from '../impl/authenticator/encodeKey';
+import keyuri from '../impl/authenticator/keyuri';
+import qrcode from '../impl/authenticator/qrcode';
+import secretKey from '../impl/authenticator/secretKey';
+import token from '../impl/authenticator/token';
 import totpCheck from '../core/totpCheck';
+import OTPUtils from './OTPUtils';
 
 /**
  * Google Authenticator adapter
@@ -35,43 +36,46 @@ import totpCheck from '../core/totpCheck';
  */
 class Authenticator {
 
-    constructor() {
-      this.opt = {
-        chart: 'https://chart.googleapis.com/chart?cht=qr&chs=150x150&choe=UTF-8&chld=M|0&chl=%uri',
-        epoch: null,
-        step: 30,
-        tokenLength: 6
-      }
+  constructor() {
+    this.opt = {
+      chart: 'https://chart.googleapis.com/chart?cht=qr&chs=150x150&choe=UTF-8&chld=M|0&chl=%uri',
+      epoch: null,
+      step: 30,
+      tokenLength: 6
     }
+  }
 
-    options(opt = {}) {
+  get utils() {
+    return OTPUtils;
+  }
 
-      // Note:
-      // only opt.chart
-      // all other options are not allowed to be overwritten
-      // since it should follow google authenticator formats.
-      this.opt.chart = opt.chart || this.opt.chart;
-    }
+  options(opt = {}) {
+    // Note:
+    // only opt.chart
+    // all other options are not allowed to be overwritten
+    // since it should follow google authenticator formats.
+    this.opt.chart = opt.chart || this.opt.chart;
+  }
 
-    encode = encodeKey
-    decode = decodeKey
-    keyuri = keyuri
+  encode = encodeKey
+  decode = decodeKey
+  keyuri = keyuri
 
-    qrcode(user, service, secret) {
-      return qrcode(user, service, secret, this.opt);
-    }
+  qrcode(user, service, secret) {
+    return qrcode(user, service, secret, this.opt);
+  }
 
-    generate(secret) {
-      return token(secret, this.opt);
-    }
+  generate(secret) {
+    return token(secret, this.opt);
+  }
 
-    generateSecret(len = 16) {
-      return secretKey(len)
-    }
+  generateSecret(len = 16) {
+    return secretKey(len)
+  }
 
-    check(token, secret){
-      return totpCheck(token, secret, this.opt);
-    }
+  check(token, secret){
+    return totpCheck(token, secret, this.opt);
+  }
 }
 
 export default Authenticator;
