+++ src/impl/authenticator/decodeKey.js
