+++ src/impl/authenticator/encodeKey.js
