+++ src/impl/authenticator/qrcode.js
