+++ src/impl/authenticator/secretKey.js
@@ -1,4 +1,4 @@
-import secretKeyUtils from '../utils/secretKey';
+import secretKeyUtils from '../../utils/secretKey';
 import encodeKey from './encodeKey';
 
 /**
